Respect `partial_rotary_factor` in config.json (Supports Phi-4-mini)
This patch starts the process of adding foundational support for Phi-4-mini by reading `partial_rotary_factor` when necessary. This is my first time making changes not to the user facing aspects of LLM hosting, but instead to the algorithms themselves. As a result, I'm having trouble determining the failure in streaming case, though the forward_(paged|tp) passes should work correctly.